examples housekeeping
