--- example/react-hello.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react'
-import { Window } from '../src'
-import { render, View, Text } from '../src/react'
-
-render(<Text>Hello</Text>, new Window("Hello"))
